requirements.txt
selenium==4.15.0
pyperclip==1.8.2
pandas==2.3.2
openpyxl==3.1.5
google-generativeai==0.8.5